Adding Magnetic Sensor Array Sketch & Script.
Adding new code for the Arduino sketch for controlling the Arduino Uno
micro-controller of the Magnetic Sensor Array device. This sketch
provides a full Hardware Abstraction Layer over the ALS31313 hall-effect
sensors, the SN74LV4051A analog multiplexers, and the electronically
activated shared I2C bus. This sketch provides the full initialization
of the hardware, and transfer of sensor measurements from the ALS31313
ICs over I2C, and out over the Serial port of the Arduino Uno to a
controlling computer. This sketch defines the message format and
protocol used for these communcations between the Arduino and computer.
This sketch is not yet completed, and requires additional
instrumentation, logging, error handling, and internal documentation.

Adding new code to implement the host side of the Magnetic Sensor Array
data recording and visualization. This is a Python script, reading the
sensor measurements either over a Serial Port, or via a formatted text
file, and displaying a real-time animation of the magnetic and
temperature fields as measured by the sensors. This script is only
partially completed, still requiring instrumentation, logging, error
handling, argument validation, and the full Serial Port logic.