Field plotting improvements.
Adding sampling rate to the display of the temperature and magnetic
field plots to show how quickly data is coming in. Adding sleeps in the
Serial reading to prevent pinning a CPU if data is not coming in fast
enough. Normalizing temperature values to the range [0,100], rather than
the full 12-bit range of the ALS31313. Adding display to the magnetic
field plot to show the average field and mean field strengths, with the
first allowing cancellation by opposing directions, and the second
purely a function of the vector norms.

Re-arranging I2CBus_t underlying Wire initialization order to actually
apply clock speed and timeout values.