Minor DRG Analysis Updates.
Adding new partially-implemeted ZStack class module to the main DRG
analysis script. Fixing improperly handled logic for running in headless
mode if a series within the *.LIF file does not contain the requested
Series. Reversing logic for z-truncation flag to default to being off
and requiring an opt-in.

Fixing Utils to use the cv2.WaitKeyEx() function instead of simple
waitKey() to get the full key-code.

Adding partial implementation of a new class for handling, working with,
and using Z-Stack images, for the DRG analysis. This is largely just
extracting out and specializing the logic originally in the
Characterize-DRGs.py script related to the Z-stacks.