Minor QOL improvements.
Updating .gitignore files to not include a handful of new local-only
scripts.

Improving the pre-processor logic in Magnetic-Probe.ino for handling
DEBUG/NDEBUG flags and the corresponding settings these result in.

Update to the requirements.txt for the Python scripts.

Improving human-readability of the timestamp value as read from the
Hall-Effect sensor in the Magnetic-Field-Sensor.py script. The
visualization now includes this, and the value is internally stored as a
number of seconds since boot.