Minor QOL improvements.
Improving window placement for multiple windows in DisplayImages().

Allowing StartFrameIndex to equal EndFrameIndex for videos of 1 frame in
duration.

Adding logic to know if images were inverted in
Rod-Alignment-Analysis.py to ensure the annotated video matches the
original video or image. Preparing to add an orientation CDF plot in
addition to the PDF plot for rod/neurite orientations. Fixing some range
issues with the orientation angles to assert they are always between
[-90,90). Adding some better visualization lines for the intermediate
stages of the elliptical filter processing.

Adding new script for creating a standardized LaTeX-formatted
experimental protocol for better documentation.