Updates to electromagnet controller firmware.
Clarifying the TWO_PHASE and THREE_PHASE macros. Adding a new stub class
for a WelfordAccumulator to compute the mean and variance of an online
stream of measurement values to separate out this logic within the
InterruptFrequency_t instance. Adding back the InterruptFrequency_t
derating factor used to account for resets caused by a too-high
interrupt frequency on the 555-Timer. Re-enabling the watchdog timer and
watchdog reset logic. Adding logic to track pin state changes. Handling
properly the case of a 0-value for the UpdatePeriod in an
InterruptFrequency_t.

Adding new scripts for uploading firmware to the Arduino controller.