Adding and debugging Welford Accumulator into Electromagnet Controller.
Refactored the online accumulators for mean and standard deviation used
in the calculation and convergence testing of the interrupt frequency
value, into a dedicated class instance. Debugged and tested out the
implementation to ensure this works interchangeably with the explicit
implementation previously used.

Adding new logging macro set to specify at compile-time the log level to
use for the firmware.

Adding to many of the outstanding ellipses with actual code comments.

Adding in the interrupt derating factor logic to handle the 555-timer
being higher frequency than can be handled by the Arduino.

Fixing implementation bug in Timestamp_t where interrupts needed to be
disabled, potentially allowing for a Timer2 Overflow event to be
skipped. Fixed with explicit checks after getting two timestamps to
ensure they differ when they must.

Making helper shell scripts for compiling and uploading smarter.