Refactoring of the Alignment-Analysis.py script.
Major refactoring of the Alignment-Analysis.py script to reduce the
number of free parameters of the elliptical filtering algorithm. Reduced
down to two defining a length scale and the image resolution. Added
logic to process either a single file or a folder of similar files.
Adding new "AlignmentMetric" value as a single encompassing metric of
alignment. Added funtionality to enable or disable the debugging image
viewing smoother.

Added Markdown How-To file for running the alignment script for other
lab members.