Refactoring of Electromagnet-Controller.ino.
Major refactoring of the Electromagnet-Controller.ino sketch to
incorporate the use of the new DRV8332 3-Phase H-Bridge controller chip
for applying the correct phase currents to the electromagnet phase
coils. This necessitated a change in interpretation of DutyCycles, with
50% representing zero current, 0% representing full negative, and 100%
full positive. Removed all traces of LoadPolarity_t, recomputed
DutyCycleLUT. Each phase is now controlled by a PWM signal which
includes polarity as noted, and an "Enable" signal to put the MOSFETs of
a given phase into a high-Z state and drive neither high nor low. Also
refactored to remove all traces of 2-phase operation.