Adding new DRG Quantification Script.
Adding the new DRG Length & Orientation Quantification script. This is
the culmination of the guidance and direction I've given to Yvonne for
this work, collected together into a testpiece to run against the best
candidate images we've collected. This is a demonstration of the
algorithms involved for optimal input data, to show that this
approach can and does work when imaging is suitable clean.

Adding README file for the new neurite quantification script.

Adding a flag to the alignment analysis script to skip pre-processing
the image, in the case where pre-processing has occurred prior to
running this script. Removing explicit references to "rods" in output
text.