Bug fixes and improved interface for DRG Batch Analysis.
Adding in status flags and return codes to DRG analysis. Adding in logic
for manual pre checking of the image files to determine whether or not a
LIF file should be processed or not. Fixing indexing in logging messages
to remove "+1" everywhere. Adding logic to allow for pre-generated DRG
body and well-interior masks. Added stub for local thresholding in
binarization of bright-field image. Adding sanity checks for DRG body
and well-interior masks. Fixed divide-by-zero opportunity in neurite
length visualization.

Fixed broken series-name/series-index logic in LIF-Splitter and ZStack.