Adding quantification results JSON file output.
Adding new classes, enumerations, and logic for reporting out the
experimental details and quantification results in a consistent JSON
file format from the DRG batch analysis. Tested and validated that all
values convert to JSON format correctly and without additional Python
class overhead. Renamed previous QuantificationResults to
QuantificationIntermediates to reflect that these are the processed Z
stacks and images, but are not the final quantification results to work
with.