Major extensions to DRG Batch Analysis results handling.
Re-factored out the DRGQuantificationResults class into the library for
working with DRG Quantification, rather than the individual analysis
script. Added new JSONDirectory argument to the scripts to allow
exporting the JSON formatted results to a batch-processing global folder
for easier further processing. Added logic to handle decoding from JSON
files and raw JSON strings. Added new columns for Phenol-Red and B27
Inclusion.

Major reworking of DRG analysis status signals and codes to be a
dedicated class to allow more enum-like handling of values.

Adding new class for handling a set of DRGQuantificationResults as the
basis for the grouping, filtering, and unique-value identification work
to be applied to the individual results.