Major updates to JSON results processing and summarization.
Major changes, improvements, bug fixes, and additions to the entire
summary figure generation pipeline. Clarified multiple variable names
and re-implemented the Additive-Exists,Additive-Concentration format to
distinguish whether a trial includes an additive and the example is a
control, versus a trial not including the additive. Added internal
methods for smoother handling of DRGQuantificationResults and
DRGQuantificationResultsSet classes. Added logic to allow filtering, and
grouping of the individual results in a ResultsSet based off either a
lambda function or a template result. Added plots to show neurite length
versus neurite density across the entire data-set, as well as GelMA
percentage and DOF, both with and without splitting on the dilution
medium and experiment date. Also added logic to export a CSV file with
the numerical data behind each figure for easy plotting in other
graphing software.

Adding logic to the batch processing script to automatically generate
the summarized result plots and figures from the JSON results once the
processing of the image files has completed.

Included caching of the ExperimentalCondition instance into the DRG
Quantification config class to allow pulling out more details during
processing when run in batch mode. Added logic to differentiate whether
the script is being run in single or batch mode. Added logic to rescale
distances in the analysis from pixels to µm, if and only if the image
resolution is known and non-zero.

Added driver script to run the re-generation of summary figures from the
JSON results by re-reading the raw data files. Added logic to allow
generating random data for the purposes of testing the plotting and data
grouping logic.