Adding new script for verifying LIF file integrity.
Adding new shell script to verify LIF file checksums to ensure they are
not trunctated or broken during transfer to and from the histone server.

Slight modifications to the DRG status code flags to remove bits when
they are no longer applicable. Adding new field and handling to
DRGQuantificationResults to tell whether a result comes from being
skipped/not imaged, or from the processing itself.