Improving Zero-Quantification results handling.
Improving and disambiguating the logic surrounding the generation of
quantification results for the DRGs which were deemed to have
"insufficient growth" and were either not imaged at all, or are not
intended for the images collected to be analyzed. Added new column to
spreadsheet, parsing and checking logic, as well as a new status code
to reflect this outcome. Added logic to handle blank or empty rows in
the spreadsheet. Improved condition handling for the cases of
SkipProcessing, InsufficientGrowth, validation failure, or pre-check
rejecton.

Fixed improperly specified analysis status codes when LIF files are
missing or have a missing series.

Fixed bug in DRG centroid estimation where the main algorithm loop would
fail immediately if no centroid was estimated with the smallest initial
kernel.

Fixed bug in GelMA Percentage handling where this was initially an
integer but needs to be a float.

Removed categories with zero data points from figure generation of DRG
results.