Improvements to DRG Quantification.
Adding new stage to the DRG quantification workflow to filter out and
remove "satellites", or clusters of pixels which are meaningfully
disconnected (by distance) from the remainder of the other pixels and
DRG body. This aims to remove the spurious dots, speckles, and ring
artefacts we're still seeing in the images, without removing features
already positively identified which are more likely to be neurites.

Also added a new check and status code to identify if and when the
neurite density computed for a DRG is concerningly high.