Minor clean-up to DRG quantification.
Cleaned up the implementation of the modification to DRG quantification
to now process the entire Z stack as individual layers for segmentation,
but then to collapse all of the identified neurites into a single 2D
image for quantification. Exporting one additional figure into the
output results and cleaned up when and where existing stacks are
collapsed into projections.