Modifications to manual interaction for DRG Quantification.
Default-enabling the manual preview of DRG images to the batch
processing script. Adding new logic to classify the images as either
acceptable or not, whether they require the ROI to be manually
specified, and/or whether there are imaging errors or not. Adding in new
status codes for the new file categorization. Adding in new handling for
more complex status codes being returned from the manual preview logic.
Sorting all conditions to assert that those requiring manual ROI
selection are processed first.